+++ admin/helpers/index.js
@@ -4,7 +4,7 @@ const moment = require('moment')
 const { version: appVersion } = require('../package.json')
 const { getBuildNumber } = require('./healthcheck')
 const config = require('../config')
-const roles = require('../lib/consts/service-roles.js')
+const roles = require('../lib/consts/roles.js')
 
 const formatPageTitle = function (pageTitle) {
   let title = 'GOV.UK'
