+++ admin/lib/consts/roles.js
@@ -1,0 +1,6 @@
+module.exports = Object.freeze({
+  helpdesk: 'HELPDESK',
+  serviceManager: 'SERVICE-MANAGER',
+  teacher: 'TEACHER',
+  testDeveloper: 'TEST-DEVELOPER'
+})
